requirements.txt
pandas==2.1.3
librosa==0.9.2
matplotlib==3.8.1
scikit-learn==1.4.2
einops==0.7.0
h5py==3.10.0
torchlibrosa==0.0.9
ftfy==6.1.1
braceexpand==0.1.7
webdataset==0.2.75
wget==3.2
timm==0.4.12
wandb==0.16.0
taming-transformers-rom1504==0.0.6
datasets==2.19.1
python-dotenv==1.0.1
transformers==4.41.0
